move auth emails to email folder
