+++ authentication/templates/authentication/email/register_parent/parent_registration_email.txt
