+++ authentication/templates/authentication/email/register_teacher/teacher_registration_email.txt
