--- authentication/templates/authentication/password-reset/password_reset_subject.txt
@@ -1,3 +1,0 @@
-{% load i18n %}{% autoescape off %}
-{% blocktrans %}Password reset on {{ site_name }}{% endblocktrans %}
-{% endautoescape %}
