+++ authentication/templates/authentication/register/register_parent_account.html
@@ -58,5 +58,4 @@
         </div>
     </div>
 </div>
-
 {% endblock content %}
