+++ authentication/templates/authentication/register/register_parent_link_existiing_login.html
@@ -66,22 +66,4 @@
         </div>
     </div>
 </div>
-
-
-<!-- <head>
-    {% load static %}
-    <link href="{% static 'authentication/style_register_parent.css' %}" rel="stylesheet">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.3/css/all.min.css" />
-</head>
-{% load crispy_forms_tags %}
-<form method="post">
-    <div class="fortschritt_balken">
-        <div class="fortschritt"></div>
-    </div>
-    <h1>{% translate 'Sign up' %}</h1>
-    <p>{% translate 'Please sign up your parent account' %}</p>
-    {% csrf_token %}
-    {{ form.email|as_crispy_field }}
-    <input type="submit" value="Submit" class="button">
-</form> -->
 {% endblock content %}
