+++ authentication/urls.py
@@ -77,12 +77,3 @@ urlpatterns = [
         name="teacher_new_registartion_view",
     ),
 ]
-
-# path('password-reset/',
-#          auth_views.PasswordResetView.as_view(
-#              template_name='authentication/password-reset/password_reset.html',
-#              subject_template_name='authentication/password-reset/password_reset_subject.txt',
-#              email_template_name='authentication/password-reset/password_reset_email.html',
-#              success_url='/login/'
-#          ),
-#          name='password_reset'),
